src/requirements.txt
aiogram==3.0.0b7
apscheduler==3.10.1
python-dotenv==1.0.0
asyncpg==0.27.0
pytz==2023.3
